Make reload take no arguments
The greeter card's reload method used to take run and tag arguments, which caused a bug because the greeter dashboard would call the card's reload method with no arguments.

We now introduce a new _fetchNewData(run, tag) method to the greeter-card component which is used as an observer and used indirectly by the card's reload method.

Fixes #11.